+++ src/environment/interval_adjustors/interval_adjustor.py
@@ -24,31 +24,3 @@ class IntervalAdjustor(metaclass=ABCMeta):
         raise NotImplementedError
 
 
-class SleepIntervalAdjustor(IntervalAdjustor):
-    def __init__(self, interval: float) -> None:
-        super().__init__(interval)
-
-    def adjust(self, offset: float = 0.0) -> float:
-        """Adjust time by time.sleep.
-
-        Args:
-            offset (float, optional): Offset value not to delay processing. Defaults to 0.0.
-
-        Returns:
-            float: Ideal value of time waited.
-        """
-        time_to_sleep = self.interval - offset
-        sleep(time_to_sleep)
-        return time_to_sleep
-
-
-class BusyLoopIntervalAdjustor(IntervalAdjustor):
-    def __init__(self, interval: float) -> None:
-        super().__init__(interval)
-
-    def adjust(self, offset: float = 0.0) -> float:
-        time_to_sleep = self.interval - offset
-        start_time = self.reset()
-        while start_time + time_to_sleep >= perf_counter():
-            pass
-        return time_to_sleep
